Added an almost-Gibbs operator for the generalized Bayesian skyline model.
The operator samples from the posterior distribution of population sizes, conditional on the genealogy,
but not including any external prior.  When no prior is provided, the Hastings ratio cancels with the
ratio of likelihoods, and the operator behaves as a Gibbs sampler.  Any external prior is taken into
account via the normal Metropolis step.

The operator can be restricted to sample only from a range of population sizes.  This is implemented
by rejection sampling.  In extreme cases, it may take -very- long to get a decent sample; so when 100
samples have been rejected, a sample at the appropriate extreme end is selected.  This is a hack.

The operator can also incorporate (without rejection sampling) a Jeffrey's prior on the geometric
average of population sizes.  This however is probably not very useful.

To make all this happen, I've also extended GammaDistribution to sample reliably all valid parameter
values, and some invalid ones (i.e. shape parameter = 0).




git-svn-id: http://beast-mcmc.googlecode.com/svn@268 ca793f91-a31e-0410-b540-2769d408b6a1