Added an additional test of full evaluation of the likelihood after an operator. This should detect operators that don't fully flag up what to update.
I also removed a likelihood.makeDirty() call that was removing efficiency savings on restore. This was presumably a work around for a bug in another class. We should expect the test of full evaluation after restore test to fail on some models (but who knows what? This call was masking the test of correct restore). Benchmark.xml works with this call removed.

git-svn-id: http://beast-mcmc.googlecode.com/svn@1558 ca793f91-a31e-0410-b540-2769d408b6a1