Lots of work on BEAUti
Started work on a modularization of BEAST generation (look in beauti.components). Lots of insertion points are defined throughout the BEAST file (relative to all the standard content) and a component can offer to write content to each or any of these. Instances of each component are created in BeautiFrame constructor. Started converting some of the monolithic generation into components (see SequenceErrorModel for an example).

Implemented the ability to import an existing BEAST file but at the moment the taxa, dates and sequences are all that are picked up.

Started work on a 'locations' tab but am currently uncertain whether this shouldn't just be a general traits tab.
 
Save/Apply templates menu options are hidden as we don't have an implementation yet.

git-svn-id: http://beast-mcmc.googlecode.com/svn@1615 ca793f91-a31e-0410-b540-2769d408b6a1