Firstly, made writeComment take a VarArg number of strings to allow for multi-line comments. Comments now have a blank line before them to break up the XML visually.
Secondly, removed the 'booleanLikelihood' for LogisticModel as this was nasty. Instead, proper priors need to be put on the t50 parameter (a gamma is given as default). Also added gamma default prior to the doubling time parameter (and picked a larger initial value to help with numerical issues in logistic and expansion model).

Thirdly, normalized the generation of discrete trait models to always use generalSubstitutionModel.

Have generated all these model and tested that they run (but not that they produce reasonable results).

git-svn-id: http://beast-mcmc.googlecode.com/svn@3404 ca793f91-a31e-0410-b540-2769d408b6a1