Adding some additional stray files that are part of the metagenome linkage model.
Adding a new model of hidden linkage among reads in a metagenomic dataset.  The model and its application to metagenomic datasets is described in a manuscript in preparation by Aaron Darling and Jonathan Eisen.  This is commit phase 1: adding new files.  Phase 2 comes next, committing related changes to existing files.  

git-svn-id: http://beast-mcmc.googlecode.com/svn@3590 ca793f91-a31e-0410-b540-2769d408b6a1