First implementation of correct calibration density for one monophyletic clade inside a Yule tree.
The internal interface it a little clunky since the concept does not fit well with the current classes implementing the various speciation priors. 

The XML interface is via an additional 'calibration' element inside
the speciationLikelihood. The element contains the taxa and a distribution for the root of the taxa. For a single taxon list, the distribution is for the parent of the taxon, which has no other constraints.
 
<speciationLikelihood id="speciationlike">
      <model>
         <yuleModel idref="yule"/>
      </model>
      <speciesTree>
          <treeModel idref="treeModel"/>
      </speciesTree>

      <calibration>
         <exponentialDistributionModel>
            <mean>
                <parameter id="mean" value="1"/>
            </mean>
         </exponentialDistributionModel>
         <taxa>
            <taxon idref="tip1"/>
            <taxon idref="tip2"/>
         </taxa>
      </calibration>
 </speciationLikelihood>


git-svn-id: http://beast-mcmc.googlecode.com/svn@3704 ca793f91-a31e-0410-b540-2769d408b6a1